+++ src/main/java/terracraft/common/init/ModBlockEntityType.java
@@ -4,18 +4,20 @@ import net.fabricmc.fabric.api.object.builder.v1.block.entity.FabricBlockEntityT
 import net.minecraft.core.Registry;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import terracraft.TerraCraft;
-import terracraft.common.entity.block.FrozenChestEntity;
-import terracraft.common.entity.block.GoldChestEntity;
-import terracraft.common.entity.block.WaterChestEntity;
+import terracraft.common.entity.block.*;
 
 public class ModBlockEntityType {
     public static final BlockEntityType<GoldChestEntity> GOLD_CHEST = FabricBlockEntityTypeBuilder.create(GoldChestEntity::new, ModBlocks.GOLD_CHEST, ModBlocks.TRAPPED_GOLD_CHEST).build(null);
     public static final BlockEntityType<FrozenChestEntity> FROZEN_CHEST = FabricBlockEntityTypeBuilder.create(FrozenChestEntity::new, ModBlocks.FROZEN_CHEST, ModBlocks.TRAPPED_FROZEN_CHEST).build(null);
+    public static final BlockEntityType<IvyChestEntity> IVY_CHEST = FabricBlockEntityTypeBuilder.create(IvyChestEntity::new, ModBlocks.IVY_CHEST, ModBlocks.TRAPPED_IVY_CHEST).build(null);
     public static final BlockEntityType<WaterChestEntity> WATER_CHEST = FabricBlockEntityTypeBuilder.create(WaterChestEntity::new, ModBlocks.WATER_CHEST).build(null);
+    public static final BlockEntityType<SkywareChestEntity> SKYWARE_CHEST = FabricBlockEntityTypeBuilder.create(SkywareChestEntity::new, ModBlocks.SKYWARE_CHEST).build(null);
 
     public static void register() {
         Registry.register(Registry.BLOCK_ENTITY_TYPE, TerraCraft.id("gold_chest"), GOLD_CHEST);
         Registry.register(Registry.BLOCK_ENTITY_TYPE, TerraCraft.id("frozen_chest"), FROZEN_CHEST);
+        Registry.register(Registry.BLOCK_ENTITY_TYPE, TerraCraft.id("ivy_chest"), IVY_CHEST);
         Registry.register(Registry.BLOCK_ENTITY_TYPE, TerraCraft.id("water_chest"), WATER_CHEST);
+        Registry.register(Registry.BLOCK_ENTITY_TYPE, TerraCraft.id("skyware_chest"), SKYWARE_CHEST);
     }
 }