+++ src/main/java/terracraft/common/init/ModLootTables.java
@@ -19,6 +19,7 @@ public class ModLootTables {
 	public static final ResourceLocation CAVE_CHEST = TerraCraft.id("chests/cave_chest");
 	public static final ResourceLocation DEEP_CAVE_CHEST = TerraCraft.id("chests/deep_cave_chest");
 	public static final ResourceLocation FROZEN_CAVE_CHEST = TerraCraft.id("chests/frozen_cave_chest");
+	public static final ResourceLocation IVY_CAVE_CHEST = TerraCraft.id("chests/ivy_cave_chest");
 
 	public static final List<ResourceLocation> INJECT_TABLE_IDS = Arrays.asList(
 			new ResourceLocation("chests/village/village_armorer"),