+++ src/main/java/terracraft/common/init/ModScreenHandler.java
@@ -8,6 +8,8 @@ public class ModScreenHandler {
     public static void register() {
         ScreenRegistry.<ChestScreenHandler, ChestBlockScreenHandler>register(ModScreenHandlerType.GOLD_CHEST, (description, inventory, title) -> new ChestBlockScreenHandler(description, inventory.player, title));
         ScreenRegistry.<ChestScreenHandler, ChestBlockScreenHandler>register(ModScreenHandlerType.FROZEN_CHEST, (description, inventory, title) -> new ChestBlockScreenHandler(description, inventory.player, title));
+        ScreenRegistry.<ChestScreenHandler, ChestBlockScreenHandler>register(ModScreenHandlerType.IVY_CHEST, (description, inventory, title) -> new ChestBlockScreenHandler(description, inventory.player, title));
         ScreenRegistry.<ChestScreenHandler, ChestBlockScreenHandler>register(ModScreenHandlerType.WATER_CHEST, (description, inventory, title) -> new ChestBlockScreenHandler(description, inventory.player, title));
+        ScreenRegistry.<ChestScreenHandler, ChestBlockScreenHandler>register(ModScreenHandlerType.SKYWARE_CHEST, (description, inventory, title) -> new ChestBlockScreenHandler(description, inventory.player, title));
     }
 }