+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.2.0] - 2022-04-06
+### Added
+- Floating Islands
+- More Biome Chests (Ivy, Sandstone, Skyware, Shadow)
+    - Ivy chests will be used for cave chests in jungle type biomes
+    - Sandstone chests will be used for cave chests in desert type biomes
+    - Skyware chests will generate on Floating Islands
+    - Shadow chests will generate in the nether
+
 ## [1.1.0] - 2022-04-06
 ### Fixed
 - Stereo audio (now mono, which makes positional)